get external avl tree working, start or connections
relays are now stored in external ram as an avl tree, allowing us to handle the large amount of hsdir relays we must keep track of in order to run an onion service. also adds or connection since I found out that tor manages connections different from circuits and it is possible to have multiple circuits on one connection. for now the connection mostly work but at the time of commit interfere with onion service tasks as they keep retreiving intro requests and for some reason hit a load/store prohibited (i don't remember which) which crashes the system. They also occasionally hang, ive noticed this happens when trying to recieve a RELAY_END after sending descriptors to an hsdir and sometimes happens after an extend call and both are likely due to a synchronization bug with the access mutex that the or connection uses.