update intro circuits and improve connection stability
intro circuits now all use the same guard relay to save time since a large portion of the time cost of a new circuit is handshaking with the first relay. also make stub init rend circuits build on exlusive connections. should probably make on the fly rend circuits do the same.

connections are more stable since a number of double free and synch bugs have been fixed. onion service was able to connect sometimes but still stalled other times. initially I thought the rend circuit was experiencing a priority inversion where the or connection task comes out of poll and for some reason poll starts running that task in stead of the rend task, stalling progress as mroe intro requests come in, so I made them run on difference cpu's so they can't interfere with each other. when that didn't stop the hanging I tried to make rend circuits on their own unique connections, though I still need to do this with on the fly rend circuits to rule out this solution, with the reasoning being that if the intro and rend circuit both use the same connection the intro requests may crowd out the relay data traffic. more testing needs to be done to see what's causing the hang.

additionally the hsdir relays picked are incorrect and don't match what the tor client calculates. i figured out that the identity is not the identity that comes in the consensus but is actually the master public key of the relay that needs to be fetched separately, which slows down consensus fetching considerably. on top of that, we are not calculating our time slice correctly and are using incorrect shared rand and time period values for relay id hashes, makes the id_hashes incorrect and in larger networks would mean our client never finds the descriptors.