got Minitor officially working on tor network
this commit was focused on getting Minitor working on the real tor network and is somewhat disorganized. a timeout was added d_recv_cell to make it wait for a maximum of 5 seconds before giving up on recving a cell. the file config.c was added to hold an array of initial relays from which we can get a consensus, on chutney they are the authorities but on real tor they are just 12 or so stable relays that I picked, in the future we should probably use the real authorities. d_fetch_descriptor_info was changed to take an array of onion rleyas, up to 3, and make a request for all their descriptors at once and fills in their master and ntor onion key. additionally relays that work as directory caches are now tracked so we can use them instead of the authorities and d_fetch_descriptor_info will keep track of which cache was fastest and use that after it checks at least 30 of the caches. d_download_consensus uses a cache relay if it has found them, which it will not have for the first fetch, and will check the consensus it already has to see if it is still valid and will skip downloading if this is the case. v_parse_s_tag had a problem where it was only parsing tags when they appeared in the order we expect in chutney and now parse regardless of what tags are actually present. v_parse_pr_tag was added to check if the relay is a dir cache. v_handle_crypto_and_insert is a new task that helps parallelize the descriptor fetch process, while the main task is fetching and waiting on a socket, this task runs the crypto functions and inserts the relays into the tree and hsdir_list. d_parse_downloaded_consensus has been changed to support this paralellization as it now creates the task and sends it the nodes to insert while it fetches new nodes from the internet. In the next commit this will be expanded so that d_parse_downloaded_consens will fetch more than 3 at a time using multiple sockets, since getting the consensus quickly is the biggest bottleneck right now. d_get_hs_time_period was created to put the code for getting the time period into one place, rather than d_push_hsdir and d_parse_downloaded_consens doing having the logic separately. relays we're changed to no longer include the issi module. since we are not currently using the qspi mode it was actually faster to put relays into a file system tree and both trees ended up being too slow so now only about 10 nodes are stored in the tree when chutney is used and 100 for real tor, the rest are put into a file hsdir_list which is just struct after struct written to the file. when we need to select which hsdir relays to store our descriptors at, we call px_get_hsdir_relays_by_id_hash which loops over every relay in the file and finds the 4 smallest 'node-idx' hash and the 4 least greater nodes, that is the 4 least node-idx values that are still bigger than our hs_index. if less than 4 greater nodes are found, the two lists are merged and the least are used as per the spec. There is still an issue where it seems Minitor and the real tor browser I'm testing with use different srv values and thus the browser is unable to find the hidden service's descriptors, when I made v_handle_crypto_and_insert only use the current srv for both id_hash and id_hash_previous it seemed to fix the issue and my browser found the same responsible hsdirs consistently until later in the day it failed again. I changed it back to how I had it originally and it started working again so this appears to be some form of timing issue. lastly some changes had to be made to hidden service to make them work on real tor, for example the payload of RELAY_CONNECTED cells had to be removed because real tor will reject any RELAY_CONNECTED cells for an onion service that has a payload.