refactor how relay descriptors are fetched
I changed the relay descriptor fetch process to use 3 connections to fetch 9 descriptors at a time, which put the hsdir processing time at 10 minutes. If we can do this simultaneously withe consensus fetch we will be able to bring the total consensus fetch down, possibly to 10 minutes total. we could also add more connections, why only 3 when we may be able to do 9?, but we will have to see if this as diminishing returns on performance. additonally all connections are made to the fastest found cache but we may want to use the fastest n connections to diversify instead of loading a single relay up with requests.