add missing rotational offset to d_get_hs_time_period and parallelize the descriptor fetching progress
after a lot of debuging I figured out that the reason Minitor wasn't connecting to clients a lot of time was because for half of the day our time period was incorrect. this is because I failed to subtract the rotation offset, which is half a day on real tor and 4 minutes on chutney, from tp_start_time, which means the in between time period and srv check was 12 hours off. this lead to Minitor only working at night, or in the last 12 hours of the day sicnce.

I also added a second task for node descritpor fetching which brought the consensus fetching process down to around 200 seconds or in the 3 to 4 minute range, 3 when you get a good fast cache. I want to change the cache selection process to make it easier to find a fast cache and thus make fast consensus fetching consistent