update local connections to pass larger chunks of data
to increase throughput without completely locking up the local connection task a large chunk of data can be read at once instead of a single relay data payload worth every pass of the loop.

I also fixed a bug where large http requests were not being recieved on the or connection. this was happening because when you read on an ssl connection it pulls all of the data out of the actual fd so I would read one cells worth, 2 cells worth would be pulled out of the fd but on the next loop iteration poll would report the fd was not readable and the 2nd cell would not be processed, now I use ioctl to calculate how much data is in the fd at the start and process that many cells worth instead of just 1 cell