overhaul Minitor to a daemon driver
Minitor now runs as a daemon and has two tasks instead of spinning up a new task for each circuit and service.

now a single connections daemon polls for both onion router connections and local connections. it also handles things like the connection creation handshake, certs exchange, challenge and netinfo cell exchange. once a connection is setup it is added to a global list of all connections and is attached to specific onion circuits with a 1 connection to many circuits relationship. when data is read in after a poll, it is put into a packed cell and sent to the core daemon as a message using a freertos queue. when connections expire they are closed, freed and removed from the list of connections by either the core daemon or the connections daemon, depending on who ran into the error.

inside of the core daemon is where services, circuits and relays are managed and used. a new hidden service is started from outside the daemon by calling d_setup_onion_service. this will then send an INIT_SERVICE message to the daemon with the service itself being the body of the message. the daemon will then call v_init_service which creates 5 INIT_CIRCUIT messages, 2 for standby that will later be used as rendezvous circuits and 3 for the introduction points of the service. these messages are sent back to the daemon and will trigger it call v_init_circuit for each one. v_init_circuit prepares a circuit and then calls d_attach_or_connection which will attach the circuit to the correct connection or create a new one if it doesn't exist. once the connection is ready, the connections daemon will send a CONN_READY message and the circuit will begin connecting and extending. the daemon knows how many times to extend the circuit by the length it was given and will stop extending once built_length is the right size. once it reaches length the daemon checks the target_status field to decide what to do next. target_status tells the daemon what a circuit's purpose is, for example CIRCUIT_ESTABLISH_INTRO means once the circuit has reached length, send and intro request. the status field of a circuit marks what state it is trying to move to, so when a circuit has status CIRCUIT_INTRO_ESTABLISHED the next cell the daemon is looking for is a RELAY_COMMAND_INTRO_ESTABLISHED cell. using status and target status, the daemon builds and extends the intro circuits and once it sees `working_circuit->service->intro_live_count == 3`, or that the service has 3 intro points ready, it calls d_push_hsdir.

d_push_hsdir is similar to the old hs desc function but instead of making a single publish circuit and looping over and over to send all the descriptors, this function just generates the descriptors as before and then makes 2 INIT_CIRCUIT messages, for the current and next time period. each of these will create, extend and then send the descriptors to the final relay, similar to the intro points before. these 2 circuits are truncated and reused to send to all the target relays. once one has finished its target relays it frees the descriptor and destroys the circuit it was using. when both have finished the hsdir timer is started to send the descriptors again later in the randomized range.

By using less tasks we save a lot of memory and make keeping track of who frees what data more manageable. the daemon can also process multiple circuits concurrently which makes it very fast compared to the old driver and if we add a second core task we could even parallelize the driver. some errors still occur with wolfssl, I plan to reach out to them to see if they want to get involved in this project.