change hsdir upload to use the sd card instead of memory
updated the hsdesc generation and upload process to save descriptors to the file system instead of memory. this allows user apps to use roughly 60k memory while the hsdesc update process is going on while only around 40k was available when memory was used. additionally sending a video file would crash the system due to the low headroom which is now no longer a problem.