add a timeout timer to restart hanging circuits
sometimes one of the relays in a circuit has problems and can take a long time to connect or may even be unreachable and never connect, while still wasting lots of our drivers time. to deal with this I added a timeout timer that periodically checks how long each circuit has gone since its last "action" and if its over 30 seconds it rebuilds the circuit.