overhaul hsdir relay tree to use file system lists instead, remove the guard marking from relays, close local sockets after 5 seconds of inactivity, fix bug in v_handle_relay_fetch related to fetches not starting when there is a waiting relay, fix bug in v_send_init_circuit_intro that led to end relays being free'd blocks, changed v_handle_conn_close and all other circuit destroy events to call v_circuit_rebuild_or_destroy instead of running individual fail sections, removed issi and and fdtree tests.
tor relays are now stored in 3 lists, 1 for hsdirs which has all hsdir nodes in the consensus, 1 for fast and stable relays to be used in regular circuits and 1 for cache relays to be used with consensus fetching. the is_guard member has been removed as it was only necessary before 1 connection could run multiple circuits.

the connections daemon now checks if a local connection has been used in the last 5 seconds and shuts it down if it has not. this is necessary becuase http sockets will be left open on the system since the client only sends a RELAY_END when it gets a zero byte response, and the connections daemon doesn't block to try and get a 0 byte read. thus in order to free the sockets up they are closed if they haven't been ready to read in 5 seconds.

v_handle_relay_fetch had a bug where a poll return would set succ to non-zero and non 1, not pdTRUE, so when the loop reset if there was a waiting_relay succ would be set to non pdTRUE, skipping the start fetch part and thus new fetches were never started. I fixed this by re-setting succ to pdFALSE and also running the start fetch part when waiting_relay is 1, regardless of the succ value

v_send_init_circuit_intro had a bug where it was checking if the current circuit we were looping over was NULL to continue the loop but it should have been checking if the end relay was set correctly. I fixed it by setting the end realy to NULL after it was freed and checkig that it was NOT NULL in the loop condition.

v_handle_conn_close now calls v_circuit_rebuild_or_destroy which handles all logic related to rebuilding, destroying and freeing a closed or malfunctioning circuit. all sections of the driver should call this instead of impelementing their own error handling logic individually.

issi is no longer used since we get good performance with just the sd card