move the connection handshake logic to the core task and lower the stack size for both the daemons
in order to lower the stack usage of the connections daemon all connection handshake logic was moved to the core daemon. This required chaning how packed cells are passed between the daemons, now or connections have a ring buffer, the connections daemon adds new cells to the ring buffer and the core daemon pops them off and processes them. the commands that come over the core queue now just inform the core task to check a connection's ring buffer instead of passing the actual cell. this was done because the connections daemon previously needed to advance the connection status to know what to do with an incoming cell, for example which cird_id length it needed to use and because the status change depends on action by the core task some kind of synchronization would have been necessary. now cells are simply put into the ring buffer for the core task to process when it gets to it, and a simple boolean is used to determine if the connections daemon should use the short or long circ_id length.

with everything moved the connections daemon now needs less stack space, more work needs to be done to see if the core daemon can be refactored to use less stack space